requirements.txt
absl-py==0.9.0
amqp==1.4.9
anyjson==0.3.3
asgiref==3.2.7
astor==0.8.0
attrs==19.3.0
Automat==20.2.0
billiard==3.3.0.23
blinker==1.4
brotlipy==0.7.0
cachetools==4.1.0
celery==3.1.26.post2
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
chromedriver-binary==2.38.0
click==7.1.2
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
Django==3.0.3
django-celery==3.3.1
django-picklefield==3.0.1
django-timezone-field==4.0
emoji==0.5.4
ephem==3.7.7.1
flower==0.9.4
future==0.18.2
gast==0.2.2
google-auth @ file:///tmp/build/80754af9/google-auth_1594357566944/work
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
h5py==2.10.0
humanize==0.5.1
hyperlink==19.0.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
importlib-metadata==1.7.0
incremental==17.5.0
itemadapter==0.1.0
jieba==0.42.1
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing==1.1.0
kombu==3.0.37
lxml==4.5.1
Markdown==3.1.1
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mysqlclient==1.4.6
numpy @ file:///C:/ci/numpy_and_numpy_base_1596233948451/work
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.1.0
parsel==1.6.0
Pillow @ file:///C:/ci/pillow_1594299032796/work
Protego==0.1.16
protobuf==3.12.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyDispatcher==2.0.5
PyHamcrest==2.0.2
PyJWT==1.7.1
PyMySQL==0.9.3
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyreadline==2.1
PySocks==1.7.1
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
queuelib==1.5.0
redis==2.10.6
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
rsa==4.0
scipy @ file:///C:/ci/scipy_1592930618155/work
Scrapy==2.2.0
scrapy-djangoitem==1.1.1
scrapyd==1.2.1
scrapyd-client==1.1.0
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
tornado==6.0.4
Twisted==20.3.0
urllib3==1.25.9
vine==1.3.0
w3lib==1.22.0
Werkzeug==0.16.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.12.1
zipp==3.1.0
zope.interface==5.1.0
